face-recognition-service/requirements.txt
# Python dependencies for face recognition service
deepface==0.0.79
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
cryptography==41.0.4
gunicorn==21.2.0
